Google colab file for object detection
Welcome to this Google Colab notebook, which serves as a practical implementation and demonstration of an advanced object detection system designed for security enhancement. This system utilizes the YOLOv6 model to identify and classify concealed weapons such as guns, knives, blades, and shurikens in x-ray images. The main objective of this notebook is to provide an interactive environment where users can test, modify, and understand the intricacies of weapon detection technology in real-world scenarios.